==> 01-etl-games.ipynb <==
## ETL del dataset `steam_games`

En esta jupyter notebook se desarrolla la extracción, transformación y carga (ETL) del conjunto de datos `steam_games.json.gz`.

### Descripción de Columnas en steam_games.gz.json

| Columna        | Descripción                                  | Ejemplo                                                         |
|----------------|----------------------------------------------|-----------------------------------------------------------------|
| publisher      | Empresa publicadora del contenido             | [Ubisoft, Dovetail Games - Trains, Degica]                      |
| genres         | Género del contenido                         | [Action, Adventure, Racing, Simulation, Strategy]               |
| app_name       | Nombre del contenido                          | [Warzone, Soundtrack, Puzzle Blocks]                            |
| title          | Título del contenido                          | [The Dream Machine: Chapter 4, Fate/EXTELLA - Sweet Room Dream, Fate/EXTELLA - Charming Bunny] |
| url            | URL de publicación del contenido              | [http://store.steampowered.com/app/761140/Lost_Summoner_Kitty/] |
| release_date   | Fecha de lanzamiento                          | [2018-01-04]                                                    |
| tags           | Etiquetas de contenido                       | [Simulation, Indie, Action, Adventure, Funny, Open World, First-Person, Sandbox, Free to Play] |
| reviews_url    | URL de reseñas del contenido                  | [http://steamcommunity.com/app/681550/reviews/?browsefilter=mostrecent&p=1] |
| specs          | Especificaciones                             | [Multi-player, Co-op, Cross-Platform Multiplayer, Downloadable Content] |
| price          | Precio del contenido                          | [4.99, 9.99, Free to Use, Free to Play]                         |
| early_access   | Acceso temprano                              | [False, True]                                                  |
| id             | Identificador único de contenido              | [761140, 643980, 670290]                                       |
| developer      | Desarrollador                                | [Kotoshiro, Secret Level SRL, Poolians.com]                    |


import gzip  # Para trabajar con archivos comprimidos en formato Gzip
import json  # Para trabajar con datos en formato JSON
import matplotlib.pyplot as plt  # Para crear gráficos y visualizaciones
import numpy as np  # Para operaciones matemáticas eficientes y manipulación de arrays
import pandas as pd  # Para el análisis y manipulación de datos en forma de DataFrames
import pyarrow as pa  # Herramientas para trabajar con datos en formato de flecha (Arrow)
import pyarrow.parquet as pq  # Para trabajar con archivos en formato Parquet
import seaborn as sns  # Biblioteca de visualización de datos basada en matplotlib

# Leemos el archivo gz que contiene el json y lo guardamos en una lista
row = []  # creamos una lista vacía para ir agregando las filas del archivo json
with gzip.open("data/steam_games.json.gz", 'rt', encoding='utf-8') as file:
    for line in file.readlines():  # leemos línea por línea
        data = json.loads(line)
        row.append(data)

# Convertimos la lista en un DataFrame
games = pd.DataFrame(row)
games.head()


games.shape # Observamos el tamaño del dataset

Ahora es necesario observar el número de nulos presentes en el dataset, para un posible imputación o eliminación de datos

games.isna().sum().sort_values(ascending= False)/len(games) * 100

Conclusión:

* Para la totalidad de columnas se presenta un valor mayor del 70% de vacíos
* No se puede concluir la eliminación de columnas ya que todos presentan alto procentaje de nulos
* Se eliminarán las filas que contengas el todas de sus registros vaciós

games = games.dropna(how="all").reset_index(drop=True) # Eliminamos las filas que tienen todos sus registros como nulos
games.head()

games.shape

games.isna().sum().sort_values(ascending= False)/len(games) * 100

Conclusión:

* Se redujo bastante el número de nulos en el dataset por columna, así como su tamaño de registros, disminuyó un 70% aproximadamente
* Publisher y developer, se pueden tolerar con los vacíos
* Genres es necesario imputar datos para evitar la perdida de información

Vamos a observar las columnas para tener una mejor visión acerca de la imputación de datos

# se observa el tipo de dato en 'genres' para varios registros de forma aleatoria
print(games['genres'].iloc[0])
print(games['genres'].iloc[15544])
print(games['genres'].iloc[30])
print(games['genres'].iloc[14])
print(games['genres'].iloc[256])

print(games['tags'].iloc[0])
print(games['tags'].iloc[15544])
print(games['tags'].iloc[30])
print(games['tags'].iloc[14])
print(games['tags'].iloc[256])

print(games['specs'].iloc[0])
print(games['specs'].iloc[15544])
print(games['specs'].iloc[30])
print(games['specs'].iloc[14])
print(games['specs'].iloc[256])

# se observa el tipo de dato en 'price'
games['price'].unique()

# se observa el tipo de dato en 'early_access'
games['early_access'].unique()

duplicados_columnas = games[games.duplicated(subset=['id'], keep=False)] # Verificamos si hay filas duplicadas en la columna id
duplicados_columnas


games[games['developer']=='Rocksteady Studios,Feral Interactive (Mac)']

# se eliminan los index deseados
id_drop = [14573, 74, 30961]
games.drop(id_drop, inplace= True)

duplicados_columnas = games[games.duplicated(subset=['id'], keep=False)] # Verificamos si hay filas duplicadas en la columna id
duplicados_columnas

games['release_date'].value_counts()

# Convertimos la columna en datetime y rellenamos los valores nulos con 'Dato no disponible'
games['release_date'] = pd.to_datetime(games['release_date'], errors='coerce')
games = games.dropna(subset=['release_date'])

# Revisamos los valores nulos
games["release_date"].isnull().sum()

Ahora se saca el año solamente de la fecha de lanzamiento

games['release_date'] = pd.to_datetime(games['release_date'], errors='coerce')
# Creamos la columna 'release_year' y usamos la propiedad .dt.year solo para las fechas válidas
games['release_year'] = games['release_date'].dt.year
# Muestra las primeras 2 filas para verificar
games.head(2)

games.drop('release_date', axis=1, inplace= True)

games['release_year'].unique()

# Definimos una función para poder reemplazar los datos string de la columna "price" por flotantes 0.0

def str_a_float(valor):

    if pd.isna(valor): # si el valor es nulo retorno 0.0
        return 0.0
    try:
        flotante = float(valor) # si el valor es un flotante retorno el mismo valor
        return flotante
    except (ValueError, TypeError): # si el valor no es un flotante retorno 0.0
        return 0.0


games["price"] = games["price"].apply(str_a_float) # aplicamos la función a la columna "price"
games['price'].dtype

# SColumnas a transformar
col_to_fill = ['publisher', 'app_name', 'title', 'developer']
# Se rellenan los nulos
df_relleno = games[col_to_fill].fillna('Sin dato disponible')
# Se borran las columnas originales y se concatenan las rellenas con todo el dataframe
games = pd.concat([games.drop(col_to_fill, axis=1), df_relleno], axis=1)
games.head()

# Usar explode para convertir las listas en filas individuales
games = games.explode('tags')
# Mostrar los valores únicos

tags_genres = [
    "Action",
    "Casual",
    "Indie",
    "Simulation",
    "Strategy",
    "Free to Play",
    "RPG",
    "Sports",
    "Adventure",
    "Racing",
    "Early Access",
    "Massively Multiplayer",
    "Animation & Modeling",
    "Video Production",
    "Utilities",
    "Web Publishing",
    "Education",
    "Software Training",
    "Design & Illustration",
    "Audio Production",
    "Photo Editing",
    "Accounting"
]

# Conservar solo las columnas en tags_drop y eliminar el resto
games = games[games['tags'].isin(tags_genres)]


games['genres'] = games['genres'].fillna(games['tags'])

games = games.explode('genres')

games = games.drop(['tags', 'specs', 'url', 'reviews_url'], axis=1)
games.columns

games = games.drop_duplicates() # Eliminamos duplicados

games.isna().sum().sort_values(ascending= False)/len(games) * 100

# Modifica todos los registros que cumplen con la condición en la columna especificada
games.loc[(games['genres'] == 'Animation &amp; Modeling'), 'genres'] = 'Animation & Modeling'

games.loc[(games['genres'] == 'Design &amp; Illustration'), 'genres'] = 'Design & Illustration'

games = games.drop_duplicates()

games.info()

games.to_csv("data/01-steam-games.csv", index=False, encoding="utf-8")

games = pd.read_csv("data/01-steam-games.csv") # leemos el archivo csv

tabla = pa.Table.from_pandas(games) # convertimos el dataframe en una tabla
pq.write_table(tabla,"data/01-steam-games.parquet") # guardamos la tabla en un archivo parquet
