==> 02-etl-reviews.ipynb <==
## ETL del dataset `steam_reviews`

En esta jupyter notebook se desarrolla la extracción, transformación y carga (ETL) del conjunto de datos `steam_reviews.json.gz`.

### Descripción de Columnas en user_reviews.gz.json

| Columna     | Descripción                                   | Ejemplo                                                                      |
|-------------|-----------------------------------------------|------------------------------------------------------------------------------|
| user_id     | Identificador único de usuario                | [76561197970982479, evcentric, maplemage]                                   |
| user_url    | URL del perfil del usuario                    | [http://steamcommunity.com/id/evcentric]                                    |
| reviews     | Revisión del usuario en formato JSON          | {'funny': '', 'posted': 'Posted September 8, 2013.', 'last_edited': '', 'item_id': '227300', 'helpful': '0 of 1 people (0%) found this review helpful', 'recommend': True, 'review': "For a simple (it's actually not all that simple but it can be!) truck driving Simulator, it is quite a fun and relaxing game. Playing on simple (or easy?) its just the basic WASD keys for driving but (if you want) the game can be much harder and realistic with having to manually change gears, much harder turning, etc. And reversing in this game is a ♥♥♥♥♥, as I imagine it would be with an actual truck. Luckily, you don't have to reverse park it but you get extra points if you do cause it is bloody hard. But this is surprisingly a nice truck driving game and I had a bit of fun with it."} |


import gzip  # Para trabajar con archivos comprimidos en formato Gzip
import json  # Para trabajar con datos en formato JSON
import matplotlib.pyplot as plt  # Para crear gráficos y visualizaciones
import numpy as np  # Para operaciones matemáticas eficientes y manipulación de arrays
import pandas as pd  # Para el análisis y manipulación de datos en forma de DataFrames
import pyarrow as pa  # Herramientas para trabajar con datos en formato de flecha (Arrow)
import pyarrow.parquet as pq  # Para trabajar con archivos en formato Parquet
import seaborn as sns  # Biblioteca de visualización de datos basada en matplotlib
import ast # Proporciona funciones para analizar y manipular el árbol de sintaxis abstracta
from datetime import datetime, timedelta

filas_review = []  # creamos una lista vacía para ir agregando las filas del archivo json
with gzip.open("data/user_reviews.json.gz", 'rt', encoding='utf-8') as file:
    for line in file.readlines():
        filas_review.append(ast.literal_eval(line))

# Se convierte en dataframe
reviews = pd.DataFrame(filas_review)
reviews

Ahora es necesario desanidar los registros que contienen listas

duplicados_columnas = reviews[reviews.duplicated(subset=['user_id'], keep=False)] # Verificamos si hay filas duplicadas en la columna id
duplicados_columnas

# Se revisan tres usuarios  de forma aleatoria

user_id = '05041129'
user_reviews = duplicados_columnas[duplicados_columnas['user_id'] == user_id]['reviews']

for review_list in user_reviews:
    for review in review_list:
        print(review['review'])
    print('=' * 100)
    
user_id = 'SuchGayMuchWow'
user_reviews = duplicados_columnas[duplicados_columnas['user_id'] == user_id]['reviews']

for review_list in user_reviews:
    for review in review_list:
        print(review['review'])
    print('=' * 100)
    
user_id = '76561198076474887'
user_reviews = duplicados_columnas[duplicados_columnas['user_id'] == user_id]['reviews']

for review_list in user_reviews:
    for review in review_list:
        print(review['review'])
    print('=' * 100)

reviews = reviews.drop_duplicates(subset='user_id', keep='first')
duplicados_columnas = reviews[reviews.duplicated(subset=['user_id'], keep=False)]
duplicados_columnas

reviews.shape

Se revisa la columna review

reviews['reviews'][0]

| Campo         | Descripción                                                                      |
|---------------|----------------------------------------------------------------------------------|
| user_id       | Identificador único para el usuario.                                             |
| user_url      | URL del perfil del usuario en streamcommunity.                                   |
| reviews       | Lista de diccionarios que contienen información sobre los reviews del usuario.  |
| - funny       | Indica si alguien puso emoticón de gracioso al review.                           |
| - posted      | Fecha de posteo del review en formato "Posted April 21, 2011".                    |
| - last_edited | Fecha de la última edición del review.                                           |
| - item_id     | Identificador único del item (juego).                                           |
| - helpful     | Estadística donde otros usuarios indican si fue útil la información.            |
| - recommend   | Booleano que indica si el usuario recomienda o no el juego.                      |
| - review      | Sentencia string con los comentarios sobre el juego.                             |


review_norm = pd.json_normalize(reviews['reviews'].dropna()) # Utilizamos el método json_normalize para normalizar los datos
review_norm.head()

Se pierde user id y user url

# Se agrega el 'user_id' y 'user_url' a las columnas separadas 
review_norm = pd.concat([reviews[['user_id', 'user_url']], review_norm], axis=1)
review_norm.head()

# Se utiliza pd.melt para transformar las columnas en filas conservando el 'user_id' y 'user_url'
review_norm = pd.melt(review_norm, id_vars=['user_id', 'user_url'], 
                       value_vars=list(range(9)),
                       value_name='reviews')
review_norm.head()

review_norm[review_norm['user_id']=='evcentric']

# Se eliminan las filas con valor None
review_norm = review_norm.dropna()
# Se verifica que solo queden el 'user_id' con la cantidad de diccionarios que le corresponde
review_norm[review_norm['user_id']=='evcentric']

# Se separan por columnas cada una de las claves de 'reviews'
reviews = review_norm['reviews'].apply(pd.Series, dtype='object')
reviews =reviews.add_prefix('reviews_')
reviews.head()

# Se une con el 'user_id' y 'user_url'
reviews = pd.concat([review_norm[['user_id', 'user_url']], reviews], axis=1)
reviews.head()

reviews.replace('', None, inplace=True)
# Mostrar las primeras filas del DataFrame
reviews.head()


reviews.info()
print('=' * 100)
print(reviews.isna().sum().sort_values(ascending= False)/len(reviews) * 100)

# Se eliminan las columnas 'reviews_funny' y 'reviews_last_edited'
reviews = reviews.drop(columns=['reviews_funny', 'reviews_last_edited'])

reviews.columns

## Transformación de la columna 'reviews_posted'

Se necesita que la fecha donde se hizo el posteo de la review este en formato `YYYY-MM-DD`, pero se encuentra como `Posted November 5, 2011.`. Por lo tanto, es necesario procesar la fecha y extraer los elementos relevantes. Se utilizará expresiones regulares para buscar y capturar los valores de año, mes y día dentro de la cadena de texto.

reviews['reviews_posted'] = reviews['reviews_posted'].str.extract(r'Posted ([\w\s\d,]+)') # Utilizamos regex para extraer la fecha
reviews['reviews_posted'] = reviews['reviews_posted'].replace({'Posted': ''}, regex=True) # Eliminamos la palabra Posted
reviews['reviews_posted'] = pd.to_datetime(reviews['reviews_posted'], errors='coerce') # Convertimos la columna reviews_posted a datetime
reviews["reviews_posted"].isnull().sum() # Contamos los valores nulos de la columna reviews_posted

print(reviews["reviews_posted"].max()) # Buscamos la fecha maxima
print(reviews["reviews_posted"].min()) # Buscamos la fecha minima

# Definimos un rango temporal adecuado para el conjunto de datos
start_date = datetime(2010, 10, 16)
end_date = datetime(2015, 12, 31)

# Imputamos fechas aleatorias a las filas con valores nulos en 'posted'
reviews.loc[reviews['reviews_posted'].isnull(), 'reviews_posted'] = np.random.choice(pd.date_range(start_date, end_date), reviews['reviews_posted'].isnull().sum())

reviews["reviews_posted"].isnull().sum() # Revisamos nuevamente la cantidad de valores nulos de la columna posted

# Extraemos el año de la columna "release_date" y creamos una nueva columna llamada "release_year"
reviews["posted_year"] = reviews["reviews_posted"].dt.year 
reviews.head(2)

reviews = reviews.drop('reviews_posted', axis=1)

reviews.isna().sum().sort_values(ascending= False)/len(reviews) * 100

reviews = reviews.dropna(subset=['reviews_review'])

reviews.isna().sum().sort_values(ascending= False)/len(reviews) * 100

reviews.to_csv("data/02-user-reviews.csv", index=False, encoding='utf-8') # Exportamos el dataframe a csv

reviews = pd.read_csv("data/02-user-reviews.csv") # leemos el archivo csv

tabla = pa.Table.from_pandas(reviews) # convertimos el dataframe en una tabla
pq.write_table(tabla,"data/02-user-reviews.parquet") # guardamos la tabla en un archivo parquet
