==> 03-etl-items.ipynb <==
## ETL del dataset `steam_items`

En esta jupyter notebook se desarrolla la extracción, transformación y carga (ETL) del conjunto de datos `steam_items.json.gz`.

### Descripción de Columnas en user_items.gz.json

| Columna  | Descripción                                | Ejemplo                                                                                        |
|----------|--------------------------------------------|-----------------------------------------------------------------------------------------------|
| user_id  | Identificador único de usuario             | [76561197970982479, evcentric, maplemage]                                                     |
| user_url | URL del perfil del usuario                 | [http://steamcommunity.com/id/evcentric]                                                      |
| items    | Items de usuario en formato JSON           | {'item_id': '273350', 'item_name': 'Evolve Stage 2', 'playtime_forever': 58, 'playtime_2weeks': 0} |

import gzip  # Para trabajar con archivos comprimidos en formato Gzip
import json  # Para trabajar con datos en formato JSON
import matplotlib.pyplot as plt  # Para crear gráficos y visualizaciones
import numpy as np  # Para operaciones matemáticas eficientes y manipulación de arrays
import pandas as pd  # Para el análisis y manipulación de datos en forma de DataFrames
import pyarrow as pa  # Herramientas para trabajar con datos en formato de flecha (Arrow)
import pyarrow.parquet as pq  # Para trabajar con archivos en formato Parquet
import seaborn as sns  # Biblioteca de visualización de datos basada en matplotlib
import ast # Proporciona funciones para analizar y manipular el árbol de sintaxis abstracta
from datetime import datetime, timedelta

filas_items = []  # creamos una lista vacía para ir agregando las filas del archivo json
with gzip.open("data/users_items.json.gz", 'rt', encoding='utf-8') as file:
    for line in file.readlines():
        filas_items.append(ast.literal_eval(line))

# Se convierte en dataframe
items = pd.DataFrame(filas_items)
items

Ahora es necesario observar el número de nulos presentes en el dataset, para un posible imputación o eliminación de datos

items.isna().sum().sort_values(ascending= False)/len(items) * 100

# Se explora la columna items
items['items'][0]

| Campo             | Descripción                                                      |
|-------------------|------------------------------------------------------------------|
| user_id           | Identificador único del usuario.                                  |
| items_count       | Número entero que indica la cantidad de juegos consumidos por el usuario.  |
| steam_id          | Número único para la plataforma Steam.                            |
| user_url          | URL del perfil del usuario.                                       |
| items             | Lista de uno o más diccionarios que describen los items consumidos por el usuario. Cada diccionario tiene las siguientes claves: |
|   - item_id       | Identificador del item (juego).                                   |
|   - item_name     | Nombre del contenido consumido (juego).                           |
|   - playtime_forever | Tiempo acumulado que el usuario ha jugado al juego.              |
|   - playtime_2weeks  | Tiempo acumulado que el usuario ha jugado al juego en las últimas dos semanas. |


# se normaliza la columna 'items'
items_norm = pd.json_normalize(filas_items, record_path=['items'], meta=['steam_id','items_count','user_id', 'user_url'] )
items_norm

duplicados = items_norm.loc[items_norm.duplicated()]
duplicados

items_norm = items_norm.drop_duplicates(keep='first')
items_norm

items_norm = items_norm.drop('playtime_2weeks', axis=1)
items_norm.columns

items_norm['playtime_forever'].value_counts()

# Filtramos y mantenemos solo las filas donde playtime_forever no es igual a 0
items_norm = items_norm[items_norm['playtime_forever'] != 0].reset_index(drop=True)

# Pasamos el tiempo de juego total a horas
items_norm['playtime_forever'] = round(items_norm['playtime_forever']/60,2)

items_norm.info()

# Eliminamos la columna 'usr_url'
items_norm = items_norm.drop(['user_url'], axis=1)

items_norm.to_csv("data/03-users-items.csv", index=False, encoding="utf-8") 

items_norm = pd.read_csv("data/03-users-items.csv") # leemos el archivo csv

tabla = pa.Table.from_pandas(items_norm) # convertimos el dataframe en una tabla de pyarrow
pq.write_table(tabla,"data/03-users-items.parquet") # guardamos la tabla en un archivo parquet
