==> 04-feature-eng.ipynb <==
# Preparación de conjuntos de datos para consultas

Importamos las librerías necesarias

import re
import ast
import numpy as np
import pandas as pd
import pyarrow as pa
import pyarrow.parquet as pq
import nltk
from textblob import TextBlob
from nltk.sentiment import SentimentIntensityAnalyzer
from datetime import datetime, timedelta

games = pd.read_parquet('data/01-steam-games.parquet')
reviews= pd.read_parquet('data/02-user-reviews.parquet')
items = pd.read_parquet('data/03-users-items.parquet')


reviews.info()

Se pide crear una nueva columna llamada 'sentiment_analysis' que reemplace a 'reviews_review' donde se realice un análisis de sentimiento de los comentarios con la siguiente escala:

* 0 si es malo,
* 1 si es neutral o esta sin review
* 2 si es positivo.

Dado que el objetivo de este proyecto es realizar una prueba de concepto, consiguiendo un producto mínimo viable, se realiza un análisis de sentimiento básico utilizando TextBlob que es una biblioteca de procesamiento de lenguaje natural (NLP) en Python. El objetivo de esta metodología es asignar un valor numérico a un texto, en este caso a los comentarios que los usuarios dejaron para un juego determinado, para representar si el sentimiento expresado en el texto es negativo, neutral o positivo. 

Esta metodología toma una revisión de texto como entrada, utiliza TextBlob para calcular la polaridad de sentimiento y luego clasifica la revisión como negativa, neutral o positiva en función de la polaridad calculada. En este caso, se consideraron las polaridades por defecto del modelo, el cuál utiliza umbrales -0.2 y 0.2, siendo polaridades negativas por debajo de -0.2, positivas por encima de 0.2 y neutrales entre medio de ambos.

def analisis_sentimiento(review):
    '''
    Realiza un análisis de sentimiento en un texto dado y devuelve un valor numérico que representa el sentimiento.

    Esta función utiliza la librería TextBlob para analizar el sentimiento en un texto dado y
    asigna un valor numérico de acuerdo a la polaridad del sentimiento.

    Parameters:
        review (str): El texto que se va a analizar para determinar su sentimiento.

    Returns:
        int: Un valor numérico que representa el sentimiento del texto:
             - 0 para sentimiento negativo.
             - 1 para sentimiento neutral o no clasificable.
             - 2 para sentimiento positivo.
    '''
    if review is None:
        return 1
    analysis = TextBlob(review)
    polarity = analysis.sentiment.polarity
    if polarity < -0.2:
        return 0  
    elif polarity > 0.2: 
        return 2 
    else:
        return 1 

reviews['sentiment_analysis'] = reviews['reviews_review'].apply(analisis_sentimiento)
reviews.head()

# Observamos los resultados del análisis de sentimientos
conteo_sentimientos = reviews['sentiment_analysis'].value_counts()
conteo_sentimientos

def revision_sentimiento(reviews, sentiments):
    """
    Muestra ejemplos de revisiones para cada categoría de análisis de sentimiento.

    Parámetros:
    - reviews (list): Lista de revisiones.
    - sentiments (list): Lista de valores de análisis de sentimiento correspondientes a cada revisión.

    La función imprime ejemplos de revisiones para cada categoría de análisis de sentimiento, 
    seleccionando aleatoriamente hasta 5 revisiones de cada categoría.
    """
    import random
    for sentiment_value in range(3):
            print(f"Para la categoría de análisis de sentimiento {sentiment_value} se tienen estos ejemplos de reviews:")
            sentiment_reviews = [review for review, sentiment in zip(reviews, sentiments) if sentiment == sentiment_value]
            
            for i, review in enumerate(random.sample(sentiment_reviews, 5), start=1):
                print(f"Review {i}: {review}")
            
            print("="*100)

revision_sentimiento(reviews['reviews_review'], reviews['sentiment_analysis'])

Se observa que no es totalmente preciso, pero para el objetivo del proyecto, es suficientemente bueno

reviews = reviews.drop(columns=['reviews_review'])
reviews.columns

## Cantidad de dinero gastado por usuario y cantidad de items consumidos

Se crea el dataframe 'df_gastos_items' que contiene por cada usuario el gasto en videojuegos y la cantidad de items consumidos. Para ello se realizan una serie de pasos creando algunos dataframe auxiliares para luego unir los resultados.

items = pd.read_parquet('data/03-users-items.parquet')

# Se extraen las columnas 'items_count', 'user_id' e 'item_id'
gasto_items = items[['items_count', 'user_id', 'item_id']]
gasto_items.head()

Se reserva el dataframe anterior y se realiza un nuevo dataframe auxiliar, pero con el precio de cada juego.

precio_juego = games[['price', 'id']]
# Elimina los duplicados
precio_juego = precio_juego.drop_duplicates(subset='id', keep='first')
# Se renombra 'id' por 'item_id' para unir mas adelante
precio_juego = precio_juego.rename(columns={'id':'item_id'})
precio_juego

Finalmente, se unen los dos dataframe auxiliares para conformar un dataframe final llamado `df_gastos_items`.

gasto_items = gasto_items.merge(precio_juego, on='item_id', how='left')
gasto_items

gasto_items.isna().sum().sort_values(ascending= False)/len(items) * 100

fill_precio = gasto_items['price'].fillna(0.0)
# Se borra la columna original y se concatena la columna rellena con todo el dataframe
gasto_items = pd.concat([gasto_items.drop('price', axis=1), fill_precio], axis=1)
gasto_items.info()

gasto_items.isna().sum().sort_values(ascending= False)/len(items) * 100

# Se elimina 'item_id'
gasto_items = gasto_items.drop('item_id', axis=1)

# Se agrupa por usuario sumando los precios de los juegos consumidos
gasto_items_usuario = gasto_items.groupby('user_id')['price'].sum().reset_index()
gasto_items_usuario

Para volver a agregar los items consumidos por cada usuario se hace un dataframe auxiliar 'df_count_items' conteniendo la cantidad de items por usuario.

conteo_items_usuario = gasto_items[['items_count', 'user_id']]

# se eliminan los duplicados
conteo_items_usuario = conteo_items_usuario.drop_duplicates(subset='user_id', keep='first')
conteo_items_usuario

gasto_items = conteo_items_usuario.merge(gasto_items_usuario, on='user_id', how='right')
gasto_items

## Playtime_forever por usuario

Se busca generar un dataframe que contenga por género de videojuego, el tiempo jugado por cada usuario con su id y url del perfil. Para ello, se realizarán algunos pasos generando algunos dataframes auxiliares.  
En primer lugar, se extrae de `df_items` las columnas 'playtime_forever', 'user_id' y 'item_id' y se reserva.

items.columns

tiempo_total_usuario_item = items[['playtime_forever', 'user_id', 'item_id']]
tiempo_total_usuario_item

Del dataframe `df_games` se extrae el 'id' del item y el género.

item_genero = games[['genres', 'id']]
# Se renombra la columna 'id' para unirla con el dataframe anterior
item_genero = item_genero.rename(columns={'id':'item_id'})
item_genero

Ahora se unen las dos tablas anteriores para obtener los generos de todos los usuarios con sus id y url.

tiempo_juego_genero = tiempo_total_usuario_item.merge(item_genero, on='item_id')
tiempo_juego_genero = tiempo_juego_genero.drop_duplicates(subset='user_id')

tiempo_juego_genero

# Se agrupa por usuario y se suma el tiempo de juego
genero_usuario_agrupado = tiempo_juego_genero.groupby(['genres', 'user_id'])['playtime_forever'].sum().reset_index()

genero_usuario_agrupado

## Tiempo de juego y año de lanzamiento

Se crea un dataframe donde se encuentre el id_item y su año  de lanzamiento con las información del usuario y sus horas de juego

anio_lanzamiento = games[['id', 'release_year']]

# Se renombra la columna 'id' para unirla con el dataframe anterior
anio_lanzamiento = anio_lanzamiento.rename(columns={'id':'item_id'})


# Se une el lanzamiento con el item de reviews
user_time_year = tiempo_juego_genero.merge(anio_lanzamiento,how='left', on='item_id')

user_time_year = user_time_year.drop_duplicates()
user_time_year

## Ranking de géneros por tiempo de juego

En este punto lo que se busca es un dataframe auxiliar que contenga el ranking de los géneros de videojuegos en función del tiempo jugados de cada uno. Se aprovecha que ya se hizo para el punto anterior un dataframe auxiliar que contiene los géneros y el tiempo jugado, por lo que se usa el mismo. Primero se agrupa por género, luego se ordena y se agrega una columna de ranking.

ranking_genero = genero_usuario_agrupado.groupby('genres')['playtime_forever'].sum().reset_index()
# Se ordena por 'playtime_horas'

ranking_genero = ranking_genero.sort_values(by='playtime_forever', ascending=False)

# Agregar una columna con la posición del ranking
ranking_genero['ranking'] = ranking_genero['playtime_forever'].rank(ascending=False).astype(int)
ranking_genero

## Cantidad de items y contenido free por desarrollador

En este punto se busca un dataframe que contenga para cada desarrollador de juegos, los items que desarrolla, el año de lanzamiento y el precio de cada uno.

precio_anual_developer = games[['price', 'release_year', 'developer', 'id']]

# Se renombra la columna 'id' para unirla con el dataframe anterior
precio_anual_developer = precio_anual_developer.rename(columns={'id':'item_id'})

# se eliminan los duplicados
items_developer = precio_anual_developer.drop_duplicates()


### Desarrolladores top según 'reviews_recommend' y 'sentiment_analysis'

sentimiento_item = reviews[['reviews_item_id','reviews_recommend','sentiment_analysis']]

# Se renombra la columna 'id' para unirla con el dataframe anterior
sentimiento_item = sentimiento_item.rename(columns={'reviews_item_id':'item_id'})


# Se une el lanzamiento con el item de reviews
top_dev = items_developer.merge(sentimiento_item,how='left', on='item_id')

# Elimino los desarrolladores que no tienen un item con review
top_dev = top_dev.dropna(subset=['reviews_recommend'])

# Elimino duplicados
top_dev = top_dev.drop_duplicates()


top_dev

## 'release_anio' en `df_reviews`

En este punto se agrega el año de lanzamiento de un juego al dataframe `df_reviews`. En primer lugar se extraen las columnas de id del juego y el año de lanzamiento del mismo, se borran los duplicados y luego se unen con el dataframe de reviews.

lanzamiento_item = games[['id', 'release_year']]

# Se renombra la columna 'id' para unirla con el dataframe anterior
lanzamiento_item = lanzamiento_item.rename(columns={'id':'reviews_item_id'})

# se eliminan los duplicados
lanzamiento_item = lanzamiento_item.drop_duplicates()

# Se une el lanzamiento con el item de reviews
reviews = reviews.merge(lanzamiento_item, on='reviews_item_id')
reviews

Se guardan los dataframes en CSV 

dfs = [reviews, gasto_items, ranking_genero, tiempo_juego_genero, items_developer, user_time_year, top_dev]
# Nombres correspondientes a cada DataFrame
names = ['02-user-reviews', '04-gasto-items', '05-ranking-genero', '06-tiempo-juego-genero', '07-items-developer', '08-user-time-year', '09-top-dev']

for df, name in zip(dfs, names):
    archivo = f'data/{name}.csv'
    df.to_csv(archivo, index=False, encoding='utf-8')

Se pasan a parquet para optimizar el espacio para el deploy

for df, name in zip(dfs, names):
    archivo = f'data/{name}.parquet'
    pq.write_table(pa.Table.from_pandas(df), archivo)
